refactor: move the folder up
