+++ src/schemas/activatorStrategy.ts
