+++ src/schemas/selectorStrategy.ts
